// src/app/modules/front-desk/settings/rooms/models/room.model.ts
export interface Room {
  CR05_NumHab: number;
  CR05_CateHab: string;
  CR05_TipoHab: string;
  CR05_CodGrp: string;
  CR05_TotCamas: number;
  CR05_NumPax: number;
  CR05_Descripcion: string;
  CR05_EstHab: string;
  CR05_Clean: string;
  CR05_Anexo: string;
  CR05_Activo: string;
  CR05_Operador: string;
}