// src/app/modules/front-desk/settings/rooms/models/room-request.model.ts
export interface RoomRequest {
  proceso: number;
  habitacion: number;
  categoria: string;
  tipo: string;
  grupo: string;
  totCamas: number;
  totPax: number;
  descripcion: string;
  estado: string;
  clean: string;
  anexo: string;
  activo: string;
  operador: string;
  respuesta: string;
}